+++ bee-network/bee-gossip/src/network/host.rs
@@ -176,7 +176,7 @@ async fn process_internal_command(internal_command: Command, swarm: &mut Swarm<S
                 warn!("Dialing peer {} failed. Cause: {}", alias!(peer_id), e);
             }
         }
-        Command::DisconnectPeer { peer_id } => hang_up(swarm, peer_id).await,
+        Command::DisconnectPeer { peer_id } => hang_up(swarm, peer_id),
         _ => {}
     }
 }
@@ -232,7 +232,7 @@ async fn dial_peer(swarm: &mut Swarm<SwarmBehaviour>, peer_id: PeerId, peerlist:
     Ok(())
 }
 
-async fn hang_up(swarm: &mut Swarm<SwarmBehaviour>, peer_id: PeerId) {
+fn hang_up(swarm: &mut Swarm<SwarmBehaviour>, peer_id: PeerId) {
     debug!("Hanging up on: {}.", alias!(peer_id));
 
     let _ = Swarm::disconnect_peer_id(swarm, peer_id);