+++ bee-tangle/src/config.rs
@@ -1,14 +1,9 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use std::num::NonZeroUsize;
-
 use serde::Deserialize;
 
 const DEFAULT_BELOW_MAX_DEPTH: u32 = 15;
-// SAFETY: initialised with a non-zero value.
-const DEFAULT_NUM_PARTITIONS: NonZeroUsize = unsafe { NonZeroUsize::new_unchecked(16) };
-const DEFAULT_MAX_EVICTION_RETRIES: usize = 10;
 
 /// A builder type for a tangle configuration.
 #[derive(Default, Deserialize, PartialEq)]
@@ -16,10 +11,6 @@ const DEFAULT_MAX_EVICTION_RETRIES: usize = 10;
 pub struct TangleConfigBuilder {
     #[serde(alias = "belowMaxDepth")]
     below_max_depth: Option<u32>,
-    #[serde(alias = "numPartitions")]
-    num_partitions: Option<NonZeroUsize>,
-    #[serde(alias = "maxEvictionRetries")]
-    max_eviction_retries: Option<usize>,
 }
 
 impl TangleConfigBuilder {
@@ -33,8 +24,6 @@ impl TangleConfigBuilder {
     pub fn finish(self) -> TangleConfig {
         TangleConfig {
             below_max_depth: self.below_max_depth.unwrap_or(DEFAULT_BELOW_MAX_DEPTH),
-            num_partitions: self.num_partitions.unwrap_or(DEFAULT_NUM_PARTITIONS),
-            max_eviction_retries: self.max_eviction_retries.unwrap_or(DEFAULT_MAX_EVICTION_RETRIES),
         }
     }
 }
@@ -43,8 +32,6 @@ impl TangleConfigBuilder {
 #[derive(Clone)]
 pub struct TangleConfig {
     below_max_depth: u32,
-    num_partitions: NonZeroUsize,
-    max_eviction_retries: usize,
 }
 
 impl TangleConfig {
@@ -56,15 +43,5 @@ impl TangleConfig {
     /// Get the value of `below_max_depth`.
     pub fn below_max_depth(&self) -> u32 {
         self.below_max_depth
-    }
-
-    /// Get the value of `num_partitions`.
-    pub fn num_partitions(&self) -> NonZeroUsize {
-        self.num_partitions
-    }
-
-    /// Get the value of `max_eviction_retries`.
-    pub fn max_eviction_retries(&self) -> usize {
-        self.max_eviction_retries
     }
 }