+++ bee-storage/bee-storage-sled/CHANGELOG.md
@@ -19,6 +19,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 0.6.0 - 2022-03-17
+
+### Added
+
+- Implementation of `InsertStrict<MessageId, MessageMetadata>` for `Storage`;
+
+### Removed
+
+- Implementation of `Insert<MessageId, MessageMetadata>` for `Storage`;
+
 ## 0.5.0 - 2022-03-11
 
 ### Added